+++ lizard_layers/models.py
@@ -7,7 +7,7 @@ from django.utils.translation import ugettext as _
 from django.contrib.gis.db import models
 from lizard_area.models import Area
 from lizard_fewsnorm.models import ParameterCache
-from lizard_measure.models import MeasuringRod
+#from lizard_measure.models import MeasuringRod
 
 class ServerMapping(models.Model):
     """
@@ -161,11 +161,17 @@ class ParameterType(models.Model):
         blank=True,
         verbose_name=_('Parameter'),
     )
-    measuring_rod = models.ForeignKey(
-        MeasuringRod,
+    # measuring_rod = models.ForeignKey(
+    #     MeasuringRod,
+    #     null=True,
+    #     blank=True,
+    #     verbose_name=_('MeasuringRod'),
+    # )
+    measuring_rod_code = models.CharField(
+        max_length=32,
         null=True,
         blank=True,
-        verbose_name=_('MeasuringRod'),
+        verbose_name=_('MeasuringRodCode'),
     )
 
     class Meta: